// src/modules/enums/enumNodeType.ts
export const enum NodeType {
  MotivationStakeholder = "Motivation:Stakeholder",
  MotivationDriver = "Motivation:Driver",
  MotivationAssesment = "Motivation:Assesment",
  MotivationGoal = "Motivation:Goal",
  MotivationOutcome = "Motivation:Outcome",
  MotivationPrinciple = "Motivation:Principle",
  MotivationRequirement = "Motivation:Requirement",
  MotivationConstraint = "Motivation:Constraint",
  MotivationMeaning = "Motivation:Meaning",
  MotivationValue = "Motivation:Value",
  StrategyResource = "Strategy:Resource",
  StrategyCapability = "Strategy:Capability",
  StrategyValueStream = "Strategy:ValueStream",
  StrategyCourseOfAction = "Strategy:CourseOfAction",
  BusinessActor = "Business:Actor",
  BusinessRole = "Business:Role",
  BusinessCollaboration = "Business:Collaboration",
  BusinessInterface = "Business:Interface",
  BusinessProcess = "Business:Process",
  BusinessFunction = "Business:Function",
  BusinessInteraction = "Business:Interaction",
  BusinessEvent = "Business:Event",
  BusinessService = "Business:Service",
  BusinessObject = "Business:Object",
  BusinessContract = "Business:Contract",
  BusinessRepresentation = "Business:Representation",
  BusinessProduct = "Business:Product",
  ApplicationComponent = "Application:Component",
  ApplicationCollaboration = "Application:Collaboration",
  ApplicationInterface = "Application:Interface",
  ApplicationFunction = "Application:Function",
  ApplicationInteraction = "Application:Interaction",
  ApplicationProcess = "Application:Process",
  ApplicationEvent = "Application:Event",
  ApplicationService = "Application:Service",
  ApplicationDataObject = "Application:DataObject",
  TechnologyNode = "Technology:Node",
  TechnologyDevice = "Technology:Device",
  TechnologySystemSoftware = "Technology:SystemSoftware",
  TechnologyCollaboration = "Technology:Collaboration",
  TechnologyInterface = "Technology:Interface",
  TechnologyPath = "Technology:Path",
  TechnologyCommunicationNetwork = "Technology:CommunicationNetwork",
  TechnologyFunction = "Technology:Function",
  TechnologyProcess = "Technology:Process",
  TechnologyInteraction = "Technology:Interaction",
  TechnologyEvent = "Technology:Event",
  TechnologyService = "Technology:Service",
  TechnologyArtifact = "Technology:Artifact",
  TechnologyEquipment = "Technology:Equipment",
  TechnologyFacility = "Technology:Facility",
  TechnologyDistributionNetwork = "Technology:DistributionNetwork",
  TechnologyMaterial = "Technology:Material",
  ImplementationAndMigrationWorkPackage = "ImplementationAndMigration:WorkPackage",
  ImplementationAndMigrationDeliverable = "ImplementationAndMigration:Deliverable",
  ImplementationAndMigrationImplementationEvent = "ImplementationAndMigration:ImplementationEvent",
  ImplementationAndMigrationPlateau = "ImplementationAndMigration:Plateau",
  ImplementationAndMigrationGap = "ImplementationAndMigration:Gap",
  CompositeGrouping = "Composite:Grouping",
  CompositeLocation = "Composite:Location"
}